UI-fixes: logout, login and navbar distance reduce.

+++ src/App.js
@@ -3,6 +3,7 @@ import { Switch, Route, Redirect } from 'react-router-dom';
 import { isEmpty } from 'lodash';
 import CustomLayout from './Layout';
 import { Login, AdminLogin } from './pages/Auth';
+import Swal from 'sweetalert2';
 import { PriceSheet } from './pages/AdminDashboard';
 import { OrderEstimate, Dashboard, Contact, ToggleTutorials, TermsOfServices } from './pages';
 
@@ -27,14 +28,31 @@ function App() {
   }, []);
 
   const handleLogout = () => {
-    if (localStorage.getItem('role')==='admin'){
-      window.location = '/admin/login';
-    }
-    else{
-      window.location = '/login';
-
-    }
-    localStorage.clear();
+    Swal.fire({
+      title: 'Confirm Logout?',
+      icon: 'question',
+      showCancelButton: true,
+      cancelButtonText: 'No',
+      confirmButtonText: 'Logout',
+      buttonsStyling: false,
+      customClass: {
+        confirmButton:
+          'w-full inline-flex justify-center border  px-4 py-2 btn  text-base font-medium text-white focus:outline-none sm:ml-3 sm:w-auto sm:text-sm custom-btn-style hover:text-red-600 hover:bg-transparent hover:border-red-600',
+        cancelButton:
+          'mt-3 w-full inline-flex justify-center hover:underline cursor-pointer px-4 py-2 text-sm font-medium text-gray-600  sm:mt-0 sm:ml-3 sm:w-auto sm:text-sm hover:underline hover:text-red-500',   
+        }
+    }).then(async (result) => {
+      if(result.isConfirmed){
+        if (localStorage.getItem('role')==='admin'){
+          window.location = '/admin/login';
+        }
+        else{
+          window.location = '/login';
+        }
+        localStorage.clear();
+      }
+    })
+    
   };
   return (
     !loading && (
